migrate to jpa repo
